requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
duckdb==1.3.0
et_xmlfile==2.0.0
fastuuid==0.12.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
litellm==1.76.1
MarkupSafe==3.0.2
matplotlib==3.7.5
multidict==6.6.4
narwhals==2.3.0
numpy==1.26.4
openai==1.102.0
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandasai==3.0.0b19
pandasai-litellm==0.0.1
pillow==10.4.0
pip==23.2.1
plotly==6.3.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==14.0.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rpds-py==0.27.1
scipy==1.10.1
seaborn==0.12.2
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlglot==25.34.1
streamlit==1.49.1
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
yarl==1.20.1
zipp==3.23.0